requirements.txt
django==3.0.4
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
drf_yasg==1.17.1
psycopg2-binary
django-silk==4.0.1
Pillow==7.0.0
fcm_django==0.3.4
django-stubs==1.5.0
django-imagekit==4.0.2
uWSGI==2.0.17.1
django_filter==2.2.0
python-decouple